Remove the need for WatchdogStorage class
